Exception handling - delete

+++ src/main/java/com/moises/store/services/UserService.java
@@ -2,8 +2,10 @@ package com.moises.store.services;
 
 import com.moises.store.entities.User;
 import com.moises.store.repositories.UserRepository;
+import com.moises.store.services.exceptions.DatabaseException;
 import com.moises.store.services.exceptions.ResourceNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -29,7 +31,14 @@ public class UserService {
     }
 
     public void delete(Long id) {
-        userRepository.deleteById(id);
+        try {
+            if (!userRepository.existsById(id)) {
+                throw new ResourceNotFoundException(id);
+            }
+            userRepository.deleteById(id);
+        } catch (DataIntegrityViolationException e) {
+            throw new DatabaseException(e.getMessage());
+        }
     }
 
     public User update(Long id, User obj) {
